fix local time for windows (10)
The old calculation did not work when in DST (daylight savings time)
The local time was also double converted with localtime().
This is only needed for Linux and OSX because the stored time is in UTC.
On windows we already have the local time.
Did not test on Windows XP or 7…